trivia-forge/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
supabase = "^2.4.3"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-cors = "^4.0.1"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
